requirements.txt
openenv-core>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
websockets>=12.0
requests>=2.31.0
openai>=1.12.0
python-dotenv